pyproject.toml
[project]
name = "autogen-swarm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "autogen-agentchat>=0.4.8.2",
    "autogen-ext>=0.4.8.2",
    "azure-core>=1.32.0",
    "azure-identity>=1.20.0",
    "lagom>=2.7.5",
    "openai>=1.65.4",
    "python-dotenv>=1.0.1",
    "tiktoken>=0.9.0",
]

[dependency-groups]
dev = [
    "pre-commit>=4.1.0",
    "pyright>=1.1.396",
    "pytest>=8.3.5",
    "pytest-async>=0.1.1",
    "pytest-cov>=6.0.0",
    "pytest-mock>=3.14.0",
    "ruff>=0.9.10",
]

[tool.ruff]
exclude = ["typings"]

[tool.ruff.lint]
select = ["E", "W", "F", "I", "TID"]

[tool.ruff.format]
docstring-code-format = true

[tool.pyright]
venvPath = "."
venv = ".venv"
exclude = ["typings", ".venv", ".vscode", ".git"]

[tool.coverage.run]
omit = [
    "*/__init__.py",
    "autogen_swarm/hosting.py",
    "autogen_swarm/protocols/*",
    "autogen_swarm/main.py",
    "autogen_swarm/tools.py"
]
